checkerboard size is now a ROS parameter

+++ ccny_rgbd_calibrate/include/ccny_rgbd_calibrate/depth_calibrator.h
@@ -40,7 +40,7 @@ class DepthCalibrator
     ros::NodeHandle nh_private_;
 
     // parameters
-    float square_size_;
+    double square_size_;
     int n_cols_;
     int n_rows_;
     int fit_window_size_;
